backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning - Facial
opencv-python==4.10.0.84
pillow==10.2.0
facenet-pytorch==2.6.0

# Machine Learning - Voz
torch==2.2.2
torchaudio==2.2.2
transformers==4.46.3
librosa==0.10.2.post1
soundfile==0.12.1
numpy==1.26.4

# Machine Learning - Texto
sentencepiece==0.2.0
protobuf==5.28.3
accelerate==1.1.1

# Base de datos
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0

# Cache
redis==5.2.0
hiredis==3.0.0

# Utilidades
python-dotenv==1.0.1
httpx==0.28.0
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging y monitoreo
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
